Program that prints the first 50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,7 +10,7 @@
 int main(void)
 {
 	int i;
-	long long fib[50];
+	long fib[50];
 
 	fib[0] = 1;
 	fib[1] = 2;
@@ -20,8 +20,8 @@ int main(void)
 		{
 			fib[i] = fib[i - 1] + fib[i - 2];
 		}
-		printf("%lld, ", fib[i]);
+		printf("%ld, ", fib[i]);
 	}
-	printf("%lld\n", fib[47] + fib[48]);
+	printf("%ld\n", fib[47] + fib[48]);
 	return (0);
 }
